requirements.txt
flask
psycopg[binary]
faker
python-dotenv
gunicorn